Include real wget to avoid busybox bugs included in alpine 3.6+
The busybox imitation wget in alpine 3.6 appears to have problems with important AWS contexts, like `wget -O - -q http://169.254.169.254/2016-09-02/meta-data/instance-id`

An effective workaround is to include the alpine real 'wget' package. This patch does that.

(Alternatively, reverting to alpine 3.5 apparently also fixes the busybox 'wget')

https://github.com/kubernetes-incubator/kube-aws/issues/1369
gliderlabs/docker-alpine#292
gliderlabs/docker-alpine#344
https://git.busybox.net/busybox/commit/?id=a6f8651911716d1d1624712eb19e4f3608767c7e
apache/incubator-openwhisk#3715

(It would be nice to have version tags in the 'awscli' repo so people could e.g. (a) revert to the working release, or (b) if this full wget breaks for someone who was counting of some property of the busybox wget.)